Check COMMIT NOTES [001] for full notes.
