+++ src/fr/upem/net/tcp/nonblocking/StringReader.java
@@ -1,4 +1,4 @@
-package fr.upem.net.tcp.nonblocking.server;
+package fr.upem.net.tcp.nonblocking;
 
 import java.nio.ByteBuffer;
